feat: Add create_user endpoint for user creation
The code changes add a new endpoint `create_user` to handle the creation of a new user. This endpoint accepts a POST request and validates the provided data. If the data is valid, a new user is created and a success message is returned. If the data is invalid, appropriate error messages are returned.

Note: The commit message follows the conventional format of "feat: <commit message>".